Moved nutritiondata to util
